requirements.txt
Django==3.2.13
psycopg2-binary==2.9.3
djangorestframework==3.13.1
python-dotenv
pytz==2022.1
djangorestframework_simplejwt
